claude-sdk-sidecar/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.39.0
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
pyyaml==6.0.2
